Kommunikation/requirements.txt
flask
opencv-python
websockets
Pillow
numpy
